\name{BREAST}
\alias{BREAST}
\title{Example data set provided with CLAG}
\description{
  Example data set provided with CLAG.
  This data set contains microarray data from breast tumor cells,
  from 20 samples and 378 genes.
  The data sets are discribed in the article (see below).
}
\value{
  The returned value is a data frame.
}
\references{
CLAG: an unsupervised non hierarchical clustering algorithm handling biological data, Linda Dib, Alessandra Carbone, BMC Bioinformatics 2012, 13:194
}
\keyword{datasets}
