\name{DIM128}
\alias{DIM128}
\title{Example data set provided with CLAG (DIM128)}
\description{
  Example data set provided with CLAG.
  This is an artificial data set with 1024
  observations of 128 variables.
  The data sets are discribed in the article (see below).
}
\value{
  The returned value is a data frame.
}
\references{
CLAG: an unsupervised non hierarchical clustering algorithm handling biological data, Linda Dib, Alessandra Carbone, BMC Bioinformatics 2012, 13:194
}
\keyword{datasets}
