\name{DIM128_subset}
\alias{DIM128_subset}
\title{Example data set provided with CLAG (DIM128 subset)}
\description{
  Example data set provided with CLAG.
  This rows of this matrix are a subset of the rows of DIM128.
  The data sets are discribed in the article (see below).
}
\value{
  The returned value is a data frame.
}
\references{
CLAG: an unsupervised non hierarchical clustering algorithm handling biological data, Linda Dib, Alessandra Carbone, BMC Bioinformatics 2012, 13:194
}
\keyword{datasets}
