\name{GLOBINE}
\alias{GLOBINE}
\title{Example data set provided with CLAG (globine)}
\description{
  Example data set provided with CLAG.
  This is a square matrix with coevolution coefficients
  (sort of correlation measure).
  The data sets are discribed in the article (see below).
}
\value{
  The returned value is a list with the data frame and row and column ids.
}
\references{
CLAG: an unsupervised non hierarchical clustering algorithm handling biological data, Linda Dib, Alessandra Carbone, BMC Bioinformatics 2012, 13:194
}
\keyword{datasets}
